+++ src/test/java/shopping/TestConfig.java
@@ -2,8 +2,8 @@ package shopping;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
-import shopping.cart.utils.currency.ExchangeRateProvider;
-import shopping.cart.utils.currency.MockExchangeRateProvider;
+import shopping.cart.service.currency.ExchangeRateProvider;
+import shopping.cart.service.currency.MockExchangeRateProvider;
 
 @TestConfiguration
 public class TestConfig {
